update-map
creacion de mapeador de habilidad

+++ src/Fifa.Core/Habilidad.cs
@@ -2,10 +2,10 @@ namespace Fifa.Core;
 
 public class Habilidad
 {
-    public int IdHabilidad { get; set; }
+    public byte IdHabilidad { get; set; }
     public string Nombre { get; set; }
     public string Descripcion { get; set; }
-    public Habilidad(int idHabilidad, string nombre, string descripcion)
+    public Habilidad(byte idHabilidad, string nombre, string descripcion)
     {
         IdHabilidad = idHabilidad;
         Nombre = nombre;
